requirements.txt
django>=4.0.1,<4.1.0
djangorestframework
pyyaml
requests
django-cors-headers